refactor: remove message module
