+++ libs/adapter/index.ts
@@ -1,23 +1,14 @@
 export * from './cheerio/cheerio-service';
-
 export * from './entity-repository/in-memory-entity-repository';
 export * from './entity-repository/mongo-db-entity-repository';
-
 export * from './event-store/event-store-db-event-store';
 export * from './event-subscription/event-store-db-event-subscription';
-
 export * from './feature-service/in-memory-feature-service';
-
 export * from './file-store/mongo-db-file-store';
 export * from './file-store/network-exception';
 export * from './file-store/s3-file-store';
 export * from './file-store/url-file-store';
-
 export * from './logging-service/mongo-db-logging-service';
 export * from './mongo-db/mongo-db-helper';
-
-export * from './nest/graphql/graphql-jwt-auth-guard';
-export * from './nest/message/message-controller';
-export * from './nest/message/message-module';
-
+export * from './nest/graphql-jwt-auth-guard';
 export * from './public-key-challenge-store/in-memory-public-key-challenge-store';
