+++ libs/vue/index.ts
@@ -1,6 +1,5 @@
 export * from './guard/auth0-guard';
 export * from './guard/metamask-guard';
-
 export * from './plugin/auth0-plugin';
 export * from './plugin/metamask-plugin';
 export * from './plugin/pl-components-plugin';
